rag: the collection registry and the search cross — the /api/rag family lands
Three route files, one family: GET/POST /api/rag/collections (the registry —
members get the doc-Brain picker with the binding matrix blanked, admins get
full rows with collectionId echoed into every binding), PUT/DELETE
/api/rag/collections/{id} (wholesale bindings replace and the delete whose
only 400 is the auto-collection refusal), and POST /api/rag/search (the
search_knowledge MCP tool's ride — agent-key or session caller, 502 with
the failure's own sentence when the index is unreachable).

The family crosses as one unit because the registry is what the search
resolves principals against.

The first live list of real rows surfaced a latent engine bug the empty dev
table had hidden: rag_collections.schema_version is the TS schema's integer
(INT4), and Rust holds the row as i64 — sqlx refuses that at row decode, a
panic. Cast to ::int8 at all four decode sites (COL_SELECT, both RETURNING
clauses, the reindex ensure sweep), the same reason id casts to ::text.

Byte-diffed on the same sessions: the admin and member GETs literally, the
full zod 400 tables (name/description bounds, the bindings array's
elements-before-length order, the search's limit/collectionIds tables), the
write cycles through BOTH runtimes (create with strip-and-nullish bindings,
bindings replace, row read-back, delete — identical modulo id/createdAt),
the 404-after-delete, the auto-refusal, the unknown-uuid no-op delete, and
live searches with real hits byte-identical (one query is un-diffable while
the workspace-activity corpus is being indexed: TS differs from itself
call-to-call on it). The two non-uuid {id} 500s are the recorded
platform-500 divergence. Audit rows mirror.

+++ api/src/retrieval/collections.rs
@@ -61,8 +61,12 @@ fn col_of(row: &sqlx::postgres::PgRow) -> RagCollection {
     }
 }
 
+// schema_version casts to int8 for the same reason id casts to text: the
+// column is the TS schema's `integer`, and Rust holds the row as i64 (the
+// type the Qdrant-shape functions take). Uncast, sqlx refuses INT4→i64 at
+// row decode — a panic the empty dev table hid until a route listed rows.
 const COL_SELECT: &str = "select id::text, name, kind, qdrant_name, description, auto, \
-                          created_by, created_at::text, schema_version";
+                          created_by, created_at::text, schema_version::int8";
 
 #[derive(Debug, Clone, PartialEq)]
 pub struct AccessBinding {
@@ -276,7 +280,7 @@ pub async fn create_collection(
     let row = sqlx::query(
         "insert into rag_collections (name, kind, qdrant_name, description, auto, embed_dim, created_by, schema_version) \
          values ($1, 'custom', $2, $3, false, $4, $5, 2) \
-         returning id::text, name, kind, qdrant_name, description, auto, created_by, created_at::text, schema_version",
+         returning id::text, name, kind, qdrant_name, description, auto, created_by, created_at::text, schema_version::int8",
     )
     .bind(input.name)
     .bind(&qdrant_name)
@@ -354,7 +358,7 @@ pub async fn ensure_personal_collection(
     let row = sqlx::query(
         "insert into rag_collections (name, kind, qdrant_name, description, auto, embed_dim, created_by, owner_user_id, schema_version) \
          values ($1, 'personal', $2, 'Your private docs, visible only to you and your personal assistant.', false, $3, $4::uuid, $4::uuid, 2) \
-         returning id::text, name, kind, qdrant_name, description, auto, created_by, created_at::text, schema_version",
+         returning id::text, name, kind, qdrant_name, description, auto, created_by, created_at::text, schema_version::int8",
     )
     .bind(opts.name.unwrap_or("My knowledge"))
     .bind(&qdrant_name)
